Refactor Octree
