+++ src/helpers.rs
@@ -1,5 +1,8 @@
 use gl::types::*;
-use std::{ffi::c_void, mem::size_of};
+use std::{
+    ffi::{c_void, CStr},
+    mem::size_of,
+};
 
 pub unsafe fn generate_atomic_counter_buffer() -> GLuint {
     let mut buffer: u32 = 0;
@@ -61,7 +64,6 @@ where
     gl::GenTextures(1, &mut texture);
     gl::BindTexture(gl::TEXTURE_BUFFER, texture);
     gl::TexBuffer(gl::TEXTURE_BUFFER, format, texture_buffer);
-    gl::BindTexture(gl::TEXTURE_BUFFER, 0);
 
     gl::BufferData(
         gl::TEXTURE_BUFFER,
@@ -73,6 +75,17 @@ where
     gl::BindBuffer(gl::TEXTURE_BUFFER, 0);
 
     (texture, texture_buffer)
+}
+
+pub unsafe fn fill_texture_buffer_with_data<T>(texture_buffer: GLuint, data: &Vec<T>) {
+    gl::BindBuffer(gl::TEXTURE_BUFFER, texture_buffer);
+    gl::BufferData(
+        gl::TEXTURE_BUFFER,
+        (size_of::<T>() * data.len()) as isize,
+        data.as_ptr() as *const c_void,
+        gl::STATIC_DRAW,
+    );
+    gl::BindBuffer(gl::TEXTURE_BUFFER, 0);
 }
 
 pub unsafe fn generate_3d_texture(size_one_dimension: u32) -> GLuint {
@@ -185,4 +198,59 @@ pub unsafe fn bind_3d_image_texture(
     format: GLenum,
 ) {
     gl::BindImageTexture(image_index, texture, 0, gl::TRUE, 0, access, format);
+}
+
+pub extern "system" fn gl_debug_output_callback(
+    source: GLenum,
+    type_: GLenum,
+    id: GLuint,
+    severity: GLenum,
+    _length: GLsizei,
+    message: *const GLchar,
+    _user_param: *mut c_void,
+) {
+    // Ignore performance errors
+    if type_ == 33360 {
+        return;
+    }
+
+    if id == 131_169 || id == 131_185 || id == 131_218 || id == 131_204 {
+        // ignore these non-significant error codes
+        return;
+    }
+
+    println!("---------------");
+    let message = unsafe { CStr::from_ptr(message).to_str().unwrap() };
+
+    println!("Debug message ({}): {}", id, message);
+    match source {
+        gl::DEBUG_SOURCE_API => println!("Source: API"),
+        gl::DEBUG_SOURCE_WINDOW_SYSTEM => println!("Source: Window System"),
+        gl::DEBUG_SOURCE_SHADER_COMPILER => println!("Source: Shader Compiler"),
+        gl::DEBUG_SOURCE_THIRD_PARTY => println!("Source: Third Party"),
+        gl::DEBUG_SOURCE_APPLICATION => println!("Source: Application"),
+        gl::DEBUG_SOURCE_OTHER => println!("Source: Other"),
+        _ => println!("Source: Unknown enum value"),
+    }
+
+    match type_ {
+        gl::DEBUG_TYPE_ERROR => println!("Type: Error"),
+        gl::DEBUG_TYPE_DEPRECATED_BEHAVIOR => println!("Type: Deprecated Behaviour"),
+        gl::DEBUG_TYPE_UNDEFINED_BEHAVIOR => println!("Type: Undefined Behaviour"),
+        gl::DEBUG_TYPE_PORTABILITY => println!("Type: Portability"),
+        gl::DEBUG_TYPE_PERFORMANCE => println!("Type: Performance"),
+        gl::DEBUG_TYPE_MARKER => println!("Type: Marker"),
+        gl::DEBUG_TYPE_PUSH_GROUP => println!("Type: Push Group"),
+        gl::DEBUG_TYPE_POP_GROUP => println!("Type: Pop Group"),
+        gl::DEBUG_TYPE_OTHER => println!("Type: Other"),
+        _ => println!("Type: Unknown enum value"),
+    }
+
+    match severity {
+        gl::DEBUG_SEVERITY_HIGH => println!("Severity: high"),
+        gl::DEBUG_SEVERITY_MEDIUM => println!("Severity: medium"),
+        gl::DEBUG_SEVERITY_LOW => println!("Severity: low"),
+        gl::DEBUG_SEVERITY_NOTIFICATION => println!("Severity: notification"),
+        _ => println!("Severity: Unknown enum value"),
+    }
 }
