+++ src/voxelization/mod.rs
@@ -1,8 +1,5 @@
 pub mod aabb;
-pub mod octree;
 pub mod visualize;
 pub mod voxelize;
-
-pub mod helpers;
 
 pub use voxelize::build_voxel_fragment_list;
