--- src/voxelization/octree/mod.rs
@@ -1,20 +1,0 @@
-pub mod build;
-pub mod common;
-pub mod helpers;
-pub mod visualize;
-
-// Octree building stages
-mod allocate_bricks;
-mod allocate_nodes;
-mod border_transfer;
-mod flag_nodes;
-mod mipmap_center;
-mod mipmap_corners;
-mod mipmap_edges;
-mod mipmap_faces;
-mod neighbour_pointers;
-mod spread_leaf_bricks;
-mod store_node_positions;
-mod write_leaf_nodes;
-
-pub use build::build_octree;
